Develop sreehari

+++ empClient/src/app/components/user-management/homepagejobview/homepagejobview.component.ts
@@ -5,29 +5,31 @@ import Swal from 'sweetalert2';
 @Component({
   selector: 'app-homepagejobview',
   templateUrl: './homepagejobview.component.html',
-  styleUrls: ['./homepagejobview.component.css']
+  styleUrls: ['./homepagejobview.component.css'],
 })
 export class HomepagejobviewComponent implements OnInit {
+  constructor(private route: Router) {}
 
-  constructor(private route:Router) { }
+  ngOnInit(): void {}
 
-  ngOnInit(): void {
+  alert() {
+    if (localStorage.getItem('accessToken')) {
+      this.route.navigate(['/job-list']);
+    } else {
+      Swal.fire({
+        title: 'Please login first..!',
+        icon: 'warning',
+        showCancelButton: true,
+        width: '400px',
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Ok',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          localStorage.clear();
+          this.route.navigate(['login']);
+        }
+      });
+    }
   }
-  alert(){
-    Swal.fire({
-      title: 'Please login first..!',
-      icon: 'warning',
-      showCancelButton: true,
-      width:'400px',
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Ok',
-    }).then((result) => {
-      if (result.isConfirmed) {
-        localStorage.clear();
-        this.route.navigate(['login']);
-      }
-    });
-  }
-
 }
